added code to find the shortest distance rider function
The function finds the distance from the rider to the restaurant and adds on the distance from the restaurant to the customer (This values is only found once at the start of the function as it will be the same no matter which rider is picked). This distance value is then compared against the shortest distance stored and if it is shorter then the relevant rider is stored in a temp variable. Once the end of the loop has been reached, the currently stored temp rider will be returned.

On the first run of the loop the first rider in the list is chosen as the shortest rider as well as calculating it's distances. This needs to be done so that the other riders can be compared.