Added functions for distance checking X & Y cords and updating rider position
Added functions for distance checking X & Y cords and updating rider position. More functionality will need to be added for when teh rider is within a certain range. This current movement system moves each rideer to their destination (restaurant or customer) after 10 cycles of the while loop. Later functionality may be added that moves the rider based on speed but for the moment this works as a means to test extra functionality rather than being it's own functionality.