Finished reworking constructors for the restaurant, rider and customer classes
Finished reworking constructors for the restaurant, rider and customer classes. House keeping was also done on these classes. Functions that are no longer needed have been removed from these. Such as setting the X and Y coordinates of a restaurant. Once set this will not chnage so there is no need to set it again at a later point.